refactor into layered constructs architecture
