+++ docs/source/constructs/application/index.rst
@@ -1,0 +1,16 @@
+Application Constructs
+======================
+
+This is the second level (**Level 2**) of abstract constructs.
+Both Level 0 and Level 1 constructs are composed to create Level 2 constructs.
+
+These constructs are meant to specialize to user-specific use case.
+This library may provide some essential application constructs, but users are
+encouraged to follow this design philosophy in their own implementations.
+You can :ref:`create your own construct <custom constructs>` to take advantage
+of the features of this library.
+
+.. toctree::
+   :maxdepth: 2
+
+   custom
