--- docs/source/constructs/assets.rst
@@ -1,5 +1,0 @@
-Assets
-======
-
-.. automodule:: orats.constructs.assets
-   :members:
