+++ docs/source/constructs/index.rst
@@ -1,8 +1,9 @@
-Higher-Level Constructs
-=======================
+Constructs
+==========
 
 .. toctree::
    :maxdepth: 2
 
-   assets
-   options
+   api/index
+   industry/index
+   application/index
