+++ docs/source/constructs/industry/assets.rst
@@ -1,0 +1,5 @@
+Assets
+======
+
+.. automodule:: orats.constructs.industry.assets
+   :members:
