+++ docs/source/constructs/industry/index.rst
@@ -1,0 +1,13 @@
+Industry Constructs
+===================
+
+This is the first level (**Level 1**) of abstract constructs.
+:ref:`Level 0 <api constructs>` constructs are composed to create Level 1 constructs.
+
+At this level, constructs are meant to correspond with industry standard concepts.
+
+.. toctree::
+   :maxdepth: 2
+
+   assets
+   options
