+++ docs/source/constructs/industry/options.rst
@@ -1,0 +1,5 @@
+Options
+=======
+
+.. automodule:: orats.constructs.industry.options
+   :members:
