--- docs/source/constructs/options.rst
@@ -1,5 +1,0 @@
-Options
-=======
-
-.. automodule:: orats.constructs.options
-   :members:
