--- docs/source/model/index.rst
@@ -1,8 +1,0 @@
-Data Structures
-===============
-
-.. toctree::
-   :maxdepth: 2
-
-   data/request
-   data/response
