+++ src/orats/constructs/industry/common.py
